refactor(Web): Change dir test for phone api
